// src/types.ts
export enum Chain {
	POLYGON = 137,
	AMOY = 80002,
}